# improve readbyte performance