Each of these 3 changes gives roughly a 2x speedup for me, taking it from 1.6s to 0.4s, vs. BufferedStreams timing of 0.3s vs. IOBuffer timing of 0.15s

Fixes https://github.com/JuliaIO/TranscodingStreams.jl/issues/150

FWIW, reading a larger chunk at a time into a buffer is usually going to be much much faster again, since it can benefit from vectorized copies at each step, but that doesn't mean we cannot attempt to microoptimize some IO functions